+++ src/main/java/io/reactivex/rxjava3/internal/jdk8/FlowableFirstStageSubscriber.java
@@ -16,8 +16,6 @@ import java.util.NoSuchElementException;
 
 import org.reactivestreams.Subscription;
 
-import io.reactivex.rxjava3.core.myBlackhole;
-
 /**
  * Signals the first element of the source via the underlying CompletableFuture,
  * signals the a default item if the upstream is empty or signals {@link NoSuchElementException}.
@@ -55,7 +53,7 @@ public final class FlowableFirstStageSubscriber<T> extends FlowableStageSubscrib
 
     @Override
     protected void afterSubscribe(Subscription s) {
-        s.request(((Integer)(1)).intValue());
+        s.request(1);
     }
 
 }
