+++ src/main/java/io/reactivex/rxjava3/internal/operators/flowable/FlowableFlatMapMaybe.java
@@ -167,7 +167,7 @@ public final class FlowableFlatMapMaybe<T, R> extends AbstractFlowableWithUpstre
 
         void innerSuccess(InnerObserver inner, R value) {
             set.delete(inner);
-            if (get() == ((Integer)(0)).intValue() && compareAndSet(0, 1)) {
+            if (get() == 0 && compareAndSet(0, 1)) {
                 boolean d = active.decrementAndGet() == 0;
                 if (requested.get() != 0) {
                     downstream.onNext(value);
